requirements.txt
Flask==2.3.2
Flask-Mail==0.9.1
Flask-Migrate==4.0.3
python-dotenv==1.0.0
Flask-SQLAlchemy==3.0.4
SQLAlchemy==2.0.20
pandas==2.1.1
openpyxl==3.1.2
Flask-WTF==1.1.1
WTForms==3.0.1
